# reduce round-trips to the sqlite thread by using execute_fetchall()