instead of issuing `execute()` and `fetchall()` as separate calls into aiosqlite, combine them into a single round-trip to the DB thread by `execute_fetchall()`.

It's unclear whether this give a speed-up in practice (at least on M1 + SSD). My measurements are likely within the noise. I suspect it might be more relevant on something like a Raspberry PI, but I'm not in a position to test that right now.

So for now, the only benefit would be slight increased clarity of intention of the code.

| test | before | after | after / before |
| --- | --- | --- | --- |
| sync empty blocks | 13.25s | 13.04s | 0.984 |
| sync full blocks | 102.24s | 102.83s | 1.0058 |
| keep-up empty blocks | 81.02s | 80.83s | 0.9977 |